Unify exporter.go and doc.go into client.go
